update rails4.1 and fix migration file

+++ db/migrate/20140416035227_create_countries.rb
@@ -2,11 +2,11 @@ class CreateCountries < ActiveRecord::Migration
   def change
     create_table :countries do |t|
       t.string :name
-      t.int :market_scale
-      t.int :product_stock
+      t.integer :market_scale
+      t.integer :product_stock
       t.string :market_segment
-      t.double :buying_prise
-      t.double :max_selling_price
+      t.float :buying_prise
+      t.float :max_selling_price
 
       t.timestamps
     end
